# search: Reduce allocations in getRevsForMatchedRepo