~1.5% of our total [allocation count](https://console.cloud.google.com/profiler;timespan=1d/frontend;type=HEAP_ALLOC/alloc_objects;focus=github.com%2Fsourcegraph%2Fsourcegraph%2Fcmd%2Ffrontend%2Finternal%2Fsearch%2Frepos.getRevsForMatchedRepo?project=sourcegraph-dev) and 0.8% of our [allocated bytes](https://console.cloud.google.com/profiler;timespan=1d/frontend;type=HEAP_ALLOC/alloc_space;focus=github.com%2Fsourcegraph%2Fsourcegraph%2Fcmd%2Ffrontend%2Finternal%2Fsearch%2Frepos.getRevsForMatchedRepo?project=sourcegraph-dev) is coming from this function, and there were a couple of fairly easy changes that can reduce the amount of allocation this function does. 

Before:
```
❯ go test -bench GetRevs -run XXX ./cmd/frontend/internal/search/repos -benchmem
goos: darwin
goarch: amd64
pkg: github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/repos
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkGetRevsForMatchedRepo/2_conflicting-16                  1522255               826.2 ns/op           314 B/op          6 allocs/op
BenchmarkGetRevsForMatchedRepo/multiple_overlapping-16            776492              1680 ns/op            1232 B/op          9 allocs/op
PASS
ok      github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/repos   3.796s
```

After:
```
❯ go test -bench GetRevs -run XXX ./cmd/frontend/internal/search/repos -benchmem
goos: darwin
goarch: amd64
pkg: github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/repos
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkGetRevsForMatchedRepo/2_conflicting-16                  1303059               953.1 ns/op           265 B/op          5 allocs/op
BenchmarkGetRevsForMatchedRepo/multiple_overlapping-16            836583              1619 ns/op             298 B/op          5 allocs/op
PASS
ok      github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/repos   3.973s
```

After Stefan's suggestion:
```
❯ go test -bench GetRevs -run XXX ./cmd/frontend/internal/search/repos -benchmem
goos: darwin
goarch: amd64
pkg: github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/repos
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkGetRevsForMatchedRepo/2_conflicting-16                  2432112               479.0 ns/op           120 B/op          3 allocs/op
BenchmarkGetRevsForMatchedRepo/multiple_overlapping-16            939200              1270 ns/op             395 B/op          5 allocs/op
PASS
ok      github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/repos   3.939s
```

<!-- Reminder: Have you updated the changelog and relevant docs (user docs, architecture diagram, etc) ? -->